refactor: supabase 폴더명 -> api로 변경
* supabase 폴더 이름 보다 api가 더 적합할듯 합니다.
* supabase는 여러 수단 중 하나일 뿐입니다. 추후 firebase가 될 수도 있고 자체 구축된 서버를 사용할 수도 있습니다.
* 이러한 변화로 부터 프로젝트 전반의 코드를 지키려면 추상적인 네이밍을 가져가는게 좋습니다.
