+++ src/stores/queries/useMenuQueries.js
@@ -1,6 +1,6 @@
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 import isEmpty from 'lodash/isEmpty';
-import { fetchAllMenus, fetchCategoryMenus, fetchLimitedMenus } from '@/supabase/menu';
+import { fetchAllMenus, fetchCategoryMenus, fetchLimitedMenus } from '@/api/menu';
 import { QUERY_KEYS } from './constants';
 
 export const useInfiniteFetchMenus = (category, limit) => {
